Last Upload
I finished my pages
There are still a few problems but I can not figure them out. I was having a problem getting my navigation at the top to be responsive, so I just made the text size smaller. I could not get the DevCon logo to get smaller.
Similar problems with the schedule page. I made the font size smaller, but I wanted to center the text inside the table and was not able to figure that out.